Complete rewrite.
This makes pear significantly more robust: no more "remonad" warnings
and special-casing! The code is significantly smaller, easier to
understand, and easier to manager.

There is one major open question: How should input be passed to macros?

At present, it is not, and macros rely on the input being named `input`.
This approach will not work for macro_rules macros as it requires
breaking hygiene, and so a better approach is required.

One potential approach is to pass the input as the first parameter to
any macro prefixed with `pear`.